Update Prime Factorization(usualWay).cpp

+++ Number Theory/Prime Factorization(usualWay).cpp
@@ -33,7 +33,7 @@ int main ()
 {
     int n, cnt ;
     cin>>n ;
-    for(int i=2; i*i<=n; i++)
+    for(int i=2; (long long) i*i<=n; i++)
     {
             if(n%i==0)
             {
